% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{append_group_vars}
\alias{append_group_vars}
\title{Append Grouping Variables}
\usage{
append_group_vars(df, val, groups)
}
\arguments{
\item{df}{A \code{skim_df}}

\item{groups}{The grouping variables of the data frame to be summarized}
}
\value{
A data frame with the grouping variables in new columns, labeled by "group_var_1", 
"group_var_2" etc.
}
\description{
Append Grouping Variables
}
