% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{arrange.skim_df}
\alias{arrange.skim_df}
\title{Use dplyr verb arrange on skim_df objects.
#' @seealso \code{\link[dplyr:arrange]{dplyr::arrange()}}}
\usage{
arrange.skim_df(.data, ...)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link[=tbl_df]{tbl_df()}}, \code{\link[dtplyr:tbl_dt]{dtplyr::tbl_dt()}} and \code{\link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi()}}.}

\item{...}{Comma separated list of unquoted variable names, or expressions
involving variable names. Use \code{\link[=desc]{desc()}} to sort a variable in descending order.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb arrange on skim_df objects.
#' @seealso \code{\link[dplyr:arrange]{dplyr::arrange()}}
}
