% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{date_max}
\alias{date_max}
\title{Get the maximum in a vector of dates}
\usage{
date_max(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the maximum in a vector of dates
}
