% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{date_median}
\alias{date_median}
\title{Get the median in a vector of dates}
\usage{
date_median(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the median in a vector of dates
}
