% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{date_min}
\alias{date_min}
\title{Get the minimum in a vector of dates}
\usage{
date_min(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the minimum in a vector of dates
}
