% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{deprecated-v1}
\alias{deprecated-v1}
\alias{skim_to_wide}
\alias{skim_to_list}
\title{Deprecated functions from skimr v1}
\usage{
skim_to_wide(.data, ...)

skim_to_list(.data, ...)
}
\arguments{
\item{.data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\value{
Either A \code{skim_df} or a \code{skim_list} object.
}
\description{
Skimr used to offer functions that combined skimming with a secondary effect,
like reshaping the data, building a list or printing the results. Some of
these behaviors are no longer necessary. \code{\link[=skim]{skim()}} always returns a wide
data frame. Others have been replaced by functions that do a single thing.
\code{\link[=partition]{partition()}} creates a list-like object from a skimmed data frame.
}
\section{Functions}{
\itemize{
\item \code{skim_to_wide}: \code{\link[=skim]{skim()}} always produces a wide data frame.

\item \code{skim_to_list}: \code{\link[=partition]{partition()}} creates a list.
}}

