% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{filter.skim_df}
\alias{filter.skim_df}
\title{Use dplyr verb filter on skim_df objects.}
\usage{
\method{filter}{skim_df}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A skim object}

\item{...}{Logical predicates defined in terms of the variables in .data.}

\item{.preserve}{When FALSE (the default) the grouping structuree is recalculated
based on the resulting data.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb filter on skim_df objects.
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}}
}
