% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{focus}
\alias{focus}
\title{Only show a subset of summary statistics after skimming}
\usage{
focus(.data, ...)
}
\arguments{
\item{.data}{A \code{skim_df} object.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\description{
This function is a variant of \verb{dplyr::select()`` designed to work with }skim_df\verb{objects. When using}focus()\verb{, }skimr\verb{metadata columns are kept, and}skimr\verb{print methods are still utilized. Otherwise, the signature and behavior is identical to}dplyr::select()`.
}
\examples{
# Compare
iris \%>\%
  skim() \%>\%
  dplyr::select(n_missing)

iris \%>\%
  skim() \%>\%
  focus(n_missing)

# This is equivalent to
iris \%>\%
  skim() \%>\%
  dplyr::select(skim_variable, skim_type, n_missing)
}
