% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{inline_hist}
\alias{inline_hist}
\title{Generate inline histogram for numeric variables}
\usage{
inline_hist(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A length-one character vector containing the histogram.
}
\description{
The character length of the histogram is controlled by the formatting
options for character vectors.
}
