% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{inline_linegraph}
\alias{inline_linegraph}
\title{Generate inline line graph for time series variables}
\usage{
inline_linegraph(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A numeric value of 0 with a name that is a character string of histogram.
}
\description{
Generate inline line graph for time series variables
}
