% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{list_lengths_max}
\alias{list_lengths_max}
\title{Get the maximum length of the lists}
\usage{
list_lengths_max(x)
}
\arguments{
\item{x}{A vector of list data}
}
\value{
Maximum length.
}
\description{
Get the maximum length of the lists
}
