% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{list_lengths_median}
\alias{list_lengths_median}
\title{Get the median length of the lists}
\usage{
list_lengths_median(x)
}
\arguments{
\item{x}{A vector of list data}
}
\value{
Median length.
}
\description{
Get the median length of the lists
}
