% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{list_lengths_min}
\alias{list_lengths_min}
\title{Get the length of the shortest list in a vector of lists}
\usage{
list_lengths_min(x)
}
\arguments{
\item{x}{A vector of list data}
}
\value{
Minimum length.
}
\description{
Get the length of the shortest list in a vector of lists
}
