% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{list_max_length}
\alias{list_max_length}
\title{Get the length of the longest list in a vector of lists}
\usage{
list_max_length(x)
}
\arguments{
\item{x}{A vector of list data}
}
\value{
Minimum length.
}
\description{
Get the length of the longest list in a vector of lists
}
