% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{max_char}
\alias{max_char}
\title{Calculate the maximum number of characters within a character vector}
\usage{
max_char(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The min of calling nchar(x).
}
\description{
Calculate the maximum number of characters within a character vector
}
