% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean_num}
\alias{mean_num}
\title{Calculate mean and create formatted value}
\usage{
mean_num(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean and the formatted value
}
\description{
Always excludes missings
}
