% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{median_num}
\alias{median_num}
\title{Calculate median and create formatted value}
\usage{
median_num(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean and the formatted value
}
\description{
Always excludes missings
}
