% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{min_char}
\alias{min_char}
\title{Calculate the minimum number of characters within a character vector}
\usage{
min_char(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The min of calling nchar(x).
}
\description{
Calculate the minimum number of characters within a character vector
}
