% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n_complete}
\alias{n_complete}
\title{Calculate complete values}
\usage{
n_complete(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The sum of non-NULL and non-NA values
}
\description{
Complete values are not missing
}
