% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n_empty}
\alias{n_empty}
\title{Calculate the number of blank values in a character vector}
\usage{
n_empty(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The number of values in the vector equal to ""
}
\description{
A "blank" is equal to "".
}
