% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n_unique}
\alias{n_unique}
\title{Calculate the number of unique elements but remove NA}
\usage{
n_unique(x)
}
\arguments{
\item{x}{A vector}
}
\value{
unique without NA.
}
\description{
Calculate the number of unique elements but remove NA
}
