% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sfc.R
\name{ncoords_sf}
\alias{ncoords_sf}
\alias{ncoords_sf}
\alias{npaths_sf}
\title{ncoords_sf - the complete count of coordinate instances
npaths_sf - complete count of paths, (a ring that is a hole is a path)}
\usage{
ncoords_sf(x)

npaths_sf(x)
}
\arguments{
\item{x}{A vector}
}
\value{
count of coordinates, all instances
}
\description{
ncoords_sf - the complete count of coordinate instances
npaths_sf - complete count of paths, (a ring that is a hole is a path)
}
