% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{pander.data.frame}
\alias{pander.data.frame}
\title{Produce \code{pander} output of a data frame}
\usage{
\method{pander}{data.frame}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a data frame}

\item{caption}{caption(string) to be shown under the table}

\item{...}{other arguments.}
}
\description{
Produce \code{pander} output of a data frame
}
