% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander.R
\name{pander}
\alias{pander}
\alias{pander.skim_df}
\alias{pander.summary_skim_df}
\title{Render data frames as markdown}
\usage{
pander(x, caption = attr(x, "caption"), ...)

pander.skim_df(x, caption = attr(x, "caption"), ...)

pander.summary_skim_df(x, ...)
}
\arguments{
\item{x}{Object to be processed}

\item{caption}{A character scalar specifying the table's caption.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Render data frames as markdown
}
\section{Functions}{
\itemize{
\item \code{pander}: Produce \code{pander} output

\item \code{pander.skim_df}: Produce \code{pander} output of a skimmed data frame

\item \code{pander.summary_skim_df}: Pander method for a summary_skim_df object
}}

\seealso{
\code{\link[pander:pander]{pander::pander()}}
}
