% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{posixct_max}
\alias{posixct_max}
\title{Get the maximum in a vector of POSIXct dates}
\usage{
posixct_max(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the maximum in a vector of POSIXct dates
}
