% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{posixct_median}
\alias{posixct_median}
\title{Get the median in a vector of POSIXct dates}
\usage{
posixct_median(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the median in a vector of POSIXct dates
}
