% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{posixct_min}
\alias{posixct_min}
\title{Get the minimum in a vector of POSIXct dates}
\usage{
posixct_min(x)
}
\arguments{
\item{x}{A vector of date data}
}
\value{
Maximum date.
}
\description{
Get the minimum in a vector of POSIXct dates
}
