% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{print.skim_vector}
\alias{print.skim_vector}
\title{Manages print for skim_vector objects.}
\usage{
\method{print}{skim_vector}(x)
}
\arguments{
\item{x}{A \code{skim_vector} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The original \code{skim_df} object.
}
\description{
Manages print for skim_vector objects.
}
