% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{quantiles_num}
\alias{quantiles_num}
\title{Calculate 25th and 75th quantiles and create formatted value}
\usage{
quantiles_num(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The 25th and 75th and formatted values
}
\description{
Always excludes missings
}
