% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{repr}
\alias{repr}
\alias{repr_text.skim_df}
\alias{repr_text.skim_list}
\alias{repr_text.one_skim_df}
\title{Skimr printing within Jupyter notebooks}
\usage{
\method{repr_text}{skim_df}(obj, ...)

\method{repr_text}{skim_list}(obj, ...)

\method{repr_text}{one_skim_df}(obj, ...)
}
\arguments{
\item{obj}{The object to \link{print} and then return the output.}

\item{...}{ignored.}
}
\value{
None. \code{invisible(NULL)}.
}
\description{
This reproduces printed results in the console. By default Jupyter kernels
render the final object in the cell. We want the version printed by
\code{skimr} instead of the data that it contains.
}
