% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sd_num}
\alias{sd_num}
\title{Calculate standard deviation and create formatted value}
\usage{
sd_num(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean and the formatted value
}
\description{
Always excludes missings
}
