% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{select.skim_df}
\alias{select.skim_df}
\title{Use dplyr verb select on skim_df objects.}
\usage{
\method{select}{skim_df}(.data, ...)
}
\arguments{
\item{.data}{A skim object}

\item{...}{One or more unquoted expressions separated by commas.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb select on skim_df objects.
}
\seealso{
\code{\link[dplyr:select]{dplyr::select()}}
}
