% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{show_formats}
\alias{show_formats}
\title{Show formatting options currently used, by data type}
\usage{
show_formats(which = NULL)
}
\arguments{
\item{which}{A character vector. One or more of the classes whose formatting
options you wish to display.}
}
\value{
A list of option-value pairs.
}
\description{
Show formatting options currently used, by data type
}
