% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_obj.R
\name{skim-obj}
\alias{skim-obj}
\alias{has_type_column}
\alias{has_variable_column}
\alias{has_skimr_attributes}
\alias{is_skim_df}
\alias{assert_is_skim_df}
\alias{is_skim_list}
\alias{assert_is_skim_list}
\alias{could_be_skim_df}
\title{Test if an object is compatible with \code{skimr}}
\usage{
has_type_column(object)

has_variable_column(object)

has_skimr_attributes(object)

is_skim_df(object)

assert_is_skim_df(object)

is_skim_list(object)

assert_is_skim_list(object)

could_be_skim_df(object)
}
\arguments{
\item{object}{Any \code{R} object.}
}
\description{
Objects within \code{skimr} are identified by a class, but they require additional
attributes and data columns for all operations to succeed. These checks help
ensure this. While they have some application externally, they are mostly
used internally.
}
\section{Functions}{
\itemize{
\item \code{has_type_column}: Does the object have the \code{skim_type} column?

\item \code{has_variable_column}: Does the object have the \code{skim_variable} column?

\item \code{has_skimr_attributes}: Does the object have the appropriate \code{skimr} attributes?

\item \code{is_skim_df}: Is the object a \code{skim_df}?

\item \code{assert_is_skim_df}: Stop if the object is not a \code{skim_df}.

\item \code{is_skim_list}: Is the object a \code{skim_list}?

\item \code{assert_is_skim_list}: Stop if the object is not a \code{skim_list}.

\item \code{could_be_skim_df}: Is this a data frame with skim_variable and skim_type columns?
}}

