% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim.grouped_df}
\alias{skim.grouped_df}
\title{Title}
\usage{
\method{skim}{grouped_df}(.data)
}
\arguments{
\item{.data}{}
}
\value{
A grouped data frame
}
\description{
Title
}
\examples{
mtcars \%>\% group_by(cyl, gear) \%>\% skim()

}
