% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skim_filter_type}
\alias{skim_filter_type}
\title{Filter a skim_df for a single type
Takes a skim_df and returns a just those rows for a single data type
(e.g "numeric", "factor").}
\usage{
skim_filter_type(x, type_selected)
}
\arguments{
\item{x}{a skim_df object}

\item{type_selected}{name of the data type to select (e.g. "numeric")}
}
\description{
Filter a skim_df for a single type
Takes a skim_df and returns a just those rows for a single data type
(e.g "numeric", "factor").
}
\examples{
  skim(iris) \%>\% skim_filter_type("numeric")
}
