% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{skim_grouped_print}
\alias{skim_grouped_print}
\title{skim_print}
\usage{
skim_grouped_print(x)
}
\arguments{
\item{x}{A \code{skim_grouped_df} object}
}
\value{
A \code{skim_df} object
}
\description{
Manages print for grouped skim objects.
}
