% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_tee}
\alias{skim_tee}
\title{Skim a data frame without modification}
\usage{
skim_tee(.data, ...)
}
\arguments{
\item{.data}{A tbl, or an object that can be coerced into a tbl.}

\item{...}{Additional options, normally used to list individual unquoted 
column names.}
}
\value{
The input data frame.
}
\description{
This prints useful summary statistics while returning the original data.
}
