% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{skim_to_wide}
\alias{skim_to_wide}
\title{Skim to a wide data frame
Deprecated}
\usage{
skim_to_wide(.data, ...)
}
\arguments{
\item{.data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\value{
A \code{skim_df} object, which also inherits the class(es) of the input
data. The result is usually a data frame or tibble.
}
\description{
Skim to a wide data frame
Deprecated
}
