% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skim_wide_formatted}
\alias{skim_wide_formatted}
\title{Create wide dataframe of formatted values in a skim_df object}
\usage{
skim_wide_formatted(x)
}
\arguments{
\item{x}{a skim_df object}
}
\description{
\code{skim_wide_formatted} converts a \code{skim_df} object into a wide data frame
with one row for each variable in the original data frame and one column for each unique 
statistic containing the formatted values.
Variables with levels (e.g. factors) are handled by appending '_1', '_2' and so on.
This function provides users with some additional flexibility for displaying or processing 
the results of skimming. 
Formatted values are always characters.
}
\examples{
skim(iris) \%>\% skim_wide_formatted() 
skim(iris) \%>\% skim_filter_type("numeric") \%>\% skim_wide_formatted() 
}
