% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{slice.skim_df}
\alias{slice.skim_df}
\title{Use dplyr verb slice on skim_df objects.}
\usage{
\method{slice}{skim_df}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A skim object}

\item{...}{Integer row calues. Provide either positive values to keep or
negative values to drop.}

\item{.preserve}{when FALSE (the default), the grouping structure is
recalculated based on the resulting data, otherwise
it is kept as is.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb slice on skim_df objects.
}
\seealso{
\code{\link[dplyr:slice]{dplyr::slice()}}
}
