% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sorted_count}
\alias{sorted_count}
\title{Create a contingency table and arrange its levels in descending order}
\usage{
sorted_count(x)
}
\arguments{
\item{x}{An object that can be interpreted as a factor (including logical)}
}
\value{
A "table" object, which will be treated as a named numeric vector
}
\description{
In case of ties, the ordering of results is alphabetical and depends upon 
the locale. 
NA is treated as a ordinary value for sorting.
}
