% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{spark_line}
\alias{spark_line}
\title{Draw a sparkline line graph with Braille characters.}
\usage{
spark_line(x)
}
\arguments{
\item{x}{A numeric vector between 0 and 1}
}
\description{
Draw a sparkline line graph with Braille characters.
}
\examples{
\dontrun{
x <- seq(0, 1, length = 10)
spark_line(x)
}
}
