% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.skimr.R
\name{summary.skim_df}
\alias{summary.skim_df}
\title{Summary function for skim_df. This is a method of the generic function \code{summary}}
\usage{
\method{summary}{skim_df}(df)
}
\arguments{
\item{df}{a skim dataframe}
}
\value{
A summary of the dataframe \code{df}
}
\description{
Summary function for skim_df. This is a method of the generic function \code{summary}
}
\examples{

a <- skim(mtcars)
summary(a)


}
