% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{to_long}
\alias{to_long}
\title{Skim results returned as a tidy long data frame with four columns:
variable, type, stat and formatted.}
\usage{
to_long(.data, ...)
}
\arguments{
\item{.data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\value{
A \code{skim_df} object, which also inherits the class(es) of the input
data. The result is usually a data frame or tibble.
}
\description{
Skim results returned as a tidy long data frame with four columns:
variable, type, stat and formatted.
}
\examples{
to_long(iris)
}
