% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{transmute.skim_df}
\alias{transmute.skim_df}
\title{Transmute a skim_df}
\usage{
transmute.skim_df(.data, ...)
}
\description{
@param ... Name-value pairs of expressions, each with length 1 or the same
length as the number of rows in the group (if using \code{\link[=group_by]{group_by()}} or in the
entire input (if not using groups). The name of each argument will be the
name of a new variable, and the value will be its corresponding value.
Use \code{NULL} value in \code{mutate} to drop a variable.  New variables overwrite
existing variables of the same name.
@return A \code{skim_df} object, which also inherits the class(es) of the input
data. The \code{variable`` and}type` variables will be included
even if not named.
In many ways, the object behaves like a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\seealso{
\code{\link[dplyr:transmute]{dplyr::transmute()}} for the function's expected behavior.
}
