% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ts_end}
\alias{ts_end}
\title{Get the finish for a time series without the frequency}
\usage{
ts_end(x)
}
\arguments{
\item{x}{A vector of ts data}
}
\value{
Finish time.
}
\description{
Get the finish for a time series without the frequency
}
