% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ts_start}
\alias{ts_start}
\title{Get the start for a time series without the frequency}
\usage{
ts_start(x)
}
\arguments{
\item{x}{A vector of ts data}
}
\value{
Finish time.
}
\description{
Get the start for a time series without the frequency
}
