% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{arrange.skim_df}
\alias{arrange.skim_df}
\title{Use dplyr verb arrange on skim_df objects.}
\usage{
\method{arrange}{skim_df}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A skim object}

\item{...}{Comma separated list of unquoted variable names, or expressions
involving variable names.}

\item{.by_group}{If TRUE, will sort first by grouping variable.
Applies to grouped data frames only.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb arrange on skim_df objects.
}
\seealso{
\code{\link[dplyr:arrange]{dplyr::arrange()}}
}
