% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skimmers.R
\name{get_skimmers}
\alias{get_skimmers}
\alias{make_skimmer}
\title{Retrieve the summary functions for a specific data type}
\usage{
get_skimmers(column, local_skimmers = new.env(), append = TRUE)

make_skimmer(type, local_skimmers, append, ...)
}
\arguments{
\item{column}{An atomic vector or list. A column from a data frame.}

\item{local_skimmers}{An environment. Used by \code{\link[=skim]{skim()}}.}

\item{append}{A logical scalar. Used by \code{\link[=skim]{skim()}}. Whether new skim functions
should be added to the existing set of functions or whether they new
functions should replace the old.}

\item{...}{A list of functions specified by:
\itemize{
\item Their name, \code{"mean"}
\item The function itself, \code{mean}
\item A call to the function with \code{.} as a dummy argument,
\code{mean(., na.rm = TRUE)}
}

These arguments are automatically \link[rlang:quo]{quoted}. They
support \link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.

The following notations are \strong{not} supported, see examples:
\itemize{
\item An anonymous function, \code{function(x) mean(x, na.rm = TRUE)}
\item An anonymous function in \pkg{purrr} notation, \code{~mean(., na.rm = TRUE)}
}}
}
\value{
A list with two entries:
\itemize{
\item \code{funs} is a \code{fun_list}. The same as \code{\link[dplyr:funs]{dplyr::funs()}}.
\item \code{type} is the type of skimmer called.
}
}
\description{
These functions are used to set the default skimming functions for a data
type.
}
\details{
When creating your own set of skimming functions, call \code{\link[=make_skimmers]{make_skimmers()}}
within a \code{\link[=get_skimmers]{get_skimmers()}}.
}
\seealso{
\code{\link[dplyr:funs]{dplyr::funs()}}
}
