% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{inline_hist}
\alias{inline_hist}
\title{Generate inline histgram for numeric variables}
\usage{
inline_hist(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A A numeric value of 0 with a name that is a character string of histogram.
}
\description{
Generate inline histgram for numeric variables
}
