% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{inline_hist}
\alias{inline_hist}
\title{Generate inline histogram for numeric variables}
\usage{
inline_hist(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A character string of histogram.
}
\description{
Generate inline histogram for numeric variables
}
