% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{inline_linegraph}
\alias{inline_linegraph}
\title{Generate inline line graph for time series variables}
\usage{
inline_linegraph(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A length-one character vector containing a line graph.
}
\description{
Sets all data to a standard length, to match character formatting. This is
twice the number of characters set in the histogram.
}
\details{
The character length of the linegraph is controlled by the formatting
options for character vectors.
}
