% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{mutate.skim_df}
\alias{mutate.skim_df}
\title{Mutate a skim_df}
\usage{
\method{mutate}{skim_df}(.data, ...)
}
\arguments{
\item{.data}{A \code{skim_df}, which behaves like a \code{tbl.}}

\item{...}{Name-value pairs of expressions, each with length 1 or the same
length as the number of rows in the group (if using \code{\link[=group_by]{group_by()}} or in the
entire input (if not using groups). The name of each argument will be the
name of a new variable, and the value will be its corresponding value.  Use
\code{NULL} value in \code{mutate} to drop a variable.  New variables overwrite
existing variables of the same name.

The arguments in \code{...} are automatically \href{rlang::quo}{quoted} and
\href{rlang::eval_tidy}{evaluated} in the context of the data frame. They
support \href{rlang::quasiquotation}{unquoting} and splicing. See
\code{vignette("programming", package = "dplyr")} for an introduction to these
concepts.}
}
\value{
A \code{skim_df} object, which also inherits the class(es) of the input
data. In many ways, the object behaves like a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
\code{\link[dplyr:mutate]{dplyr::mutate()}} currently drops attributes, but we need to keep them around
for other skim behaviors. Otherwise the behavior is exactly the same. For
more information, see https://github.com/tidyverse/dplyr/issues/3429.
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}} for the function's expected behavior.
}
