% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{mutate.skim_df}
\alias{mutate.skim_df}
\title{Use dplyr verb mutate on skim_df objects.}
\usage{
\method{mutate}{skim_df}(.data, ...)
}
\arguments{
\item{.data}{A skim object}

\item{...}{Name-value pairs of expressions. The name of each argument will
be the name of a new variable, and the value will be its
corresponding value.}
}
\value{
skim_df object coerced to a data frame.
}
\description{
Use dplyr verb mutate on skim_df objects.
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}}
}
