% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{pander}
\alias{pander}
\title{Create pander object}
\usage{
pander(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{an R object (typically a matrix or data frame)}

\item{caption}{caption(string) to be shown under the table}

\item{...}{other arguments.}
}
\description{
Generic method for \code{pander} objects based on the method in the pander package.
}
\seealso{
\code{\link[pander]{pander}}
}
