% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander.R
\name{pander}
\alias{pander}
\alias{pander.skim_df}
\alias{pander.summary_skim_df}
\title{Render data frames as markdown}
\usage{
\method{pander}{skim_df}(x, caption = attr(x, "caption"), ...)

\method{pander}{summary_skim_df}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{caption}{A character scalar specifying the table's caption.}

\item{...}{optional parameters passed to special methods and/or raw \code{pandoc.*} functions}
}
\description{
Render data frames as markdown
}
\section{Methods (by class)}{
\itemize{
\item \code{skim_df}: Produce \code{pander} output of a skimmed data frame

\item \code{summary_skim_df}: Pander method for a summary_skim_df object
}}

\seealso{
\code{\link[pander:pander]{pander::pander()}}
}
