% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{print}
\alias{print}
\alias{print.skim_df}
\alias{print.skim_vector}
\alias{print.summary_skim_df}
\title{Print \code{skim} objects}
\usage{
\method{print}{skim_df}(x, ...)

\method{print}{skim_vector}(x, ...)

\method{print}{summary_skim_df}(x, ...)
}
\arguments{
\item{x}{Either a \code{skim_df}, \code{skim_vector} or \code{skim_summary} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print \code{skim} objects
}
\section{Methods (by class)}{
\itemize{
\item \code{skim_df}: Prints a skimmed data frame (\code{skim_df} from \code{\link[=skim]{skim()}})

\item \code{skim_vector}: Manages print for \code{skim_vector} objects.

\item \code{summary_skim_df}: Print method for a \code{summary_skim_df} object.
}}

