% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{print}
\alias{print}
\alias{print.skim_df}
\alias{print.one_skim_df}
\alias{print.skim_list}
\alias{print.summary_skim_df}
\alias{knit_print.skim_list}
\alias{knit_print.one_skim_df}
\title{Print \code{skim} objects}
\usage{
\method{print}{skim_df}(x, include_summary = TRUE, ...)

\method{print}{one_skim_df}(x, ..)

\method{print}{skim_list}(x, ..)

\method{print}{summary_skim_df}(x, ...)

\method{knit_print}{skim_list}(x, options = NULL, ...)

\method{knit_print}{one_skim_df}(x, options = NULL, ...)
}
\arguments{
\item{x}{Either a \code{skim_df}, \code{skim_vector} or \code{skim_summary} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{skimr} has custom print methods for all supported objects. Default printing
methods for \code{knitr}/ \code{rmarkdown} documents is also provided.
}
\section{Methods (by class)}{
\itemize{
\item \code{skim_df}: Print a skimmed data frame (\code{skim_df} from \code{\link[=skim]{skim()}}).

\item \code{one_skim_df}: Print an entry within a partitioned \code{skim_df}.

\item \code{skim_list}: Print a \code{skim_list}, a list of \code{skim_df} objects.

\item \code{summary_skim_df}: Print method for a \code{summary_skim_df} object.

\item \code{skim_list}: Default \code{knitr} print for a \code{skim_list}.

\item \code{one_skim_df}: Default \code{knitr} print within a partitioned \code{skim_df}.
}}

