% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_print.R
\name{print.skim_df}
\alias{print.skim_df}
\title{Print skimmed data frame}
\usage{
\method{print}{skim_df}(x, ...)
}
\arguments{
\item{x}{A \code{skim_df} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The original \code{skim_df} object.
}
\description{
Prints a skimmed data frame (created by skim()).
}
