% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.skimr.R
\name{print.summary_skim_df}
\alias{print.summary_skim_df}
\title{Print method for a summary_skim_df object. This is a method for the generic function \code{print}}
\usage{
\method{print}{summary_skim_df}(skim_summary, ...)
}
\arguments{
\item{skim_summary}{a skim_summary object}
}
\description{
Print method for a summary_skim_df object. This is a method for the generic function \code{print}
}
