% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{contains}
\alias{select_helpers}
\alias{reexports}
\alias{ends_with}
\alias{reexports}
\alias{everything}
\alias{reexports}
\alias{matches}
\alias{reexports}
\alias{num_range}
\alias{reexports}
\alias{one_of}
\alias{reexports}
\alias{starts_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tidyselect}{\code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

