% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}

   \item{tidyselect}{\code{\link[tidyselect:contains]{contains}}, \code{\link[tidyselect:ends_with]{ends_with}}, \code{\link[tidyselect:everything]{everything}}, \code{\link[tidyselect:matches]{matches}}, \code{\link[tidyselect:num_range]{num_range}}, \code{\link[tidyselect:one_of]{one_of}}, \code{\link[tidyselect:starts_with]{starts_with}}}
}}

