% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{select.skim_df}
\alias{select.skim_df}
\title{Use dplyr verb select on skim_df objects.
Always includes variable and type columns even if user has not selected them.}
\usage{
select.skim_df(.data, ...)
}
\arguments{
\item{.data}{a skim_df object}

\item{...}{list of variables to be retained}
}
\value{
skim_df object
}
\description{
Use dplyr verb select on skim_df objects.
Always includes variable and type columns even if user has not selected them.
}
\seealso{
\code{\link[dplyr:select]{dplyr::select()}}
}
