% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_skimmers}
\alias{show_skimmers}
\title{Show summary functions currently used, by column type}
\usage{
show_skimmers()
}
\value{
Nothing. \code{invisible()}
}
\description{
Show summary functions currently used, by column type
}
