% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_skimmers}
\alias{show_skimmers}
\title{Show summary functions currently used, by data type}
\usage{
show_skimmers(which = NULL)
}
\arguments{
\item{which}{A character vector. One or more of the classes whose summary
functions you wish to display.}
}
\value{
A list. The names of the list match the classes that have assigned
 summary functions.
}
\description{
The names of the summary functions are stored as a character vector within
a list. The names of the values in the list match the names of the classes
that have assigned sets of summary functions.
}
