% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim.grouped_df}
\alias{skim.grouped_df}
\title{Skim function for grouped date}
\usage{
\method{skim}{grouped_df}(.data)
}
\arguments{
\item{.data}{A grouped tbl, or an object that can be coerced into a tbl.}
}
\value{
A grouped skim_df data frame
}
\description{
Skim function for grouped date
}
\examples{
mtcars \%>\% dplyr::group_by(cyl, gear) \%>\% skim()

}
