% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_tee}
\alias{skim_tee}
<<<<<<< HEAD
\title{Print useful summary statistic from a data frame without modification}
=======
\title{Print useful summary statistic from a data frame returning the data frame
 without modification}
>>>>>>> upstream/develop
\usage{
skim_tee(.data, ...)
}
\arguments{
\item{.data}{A tbl, or an object that can be coerced into a tbl.}

\item{...}{Additional options, normally used to list individual unquoted column names.}
}
\value{
The input data frame.
}
\description{
<<<<<<< HEAD
Print useful summary statistic from a data frame without modification
=======
Print useful summary statistic from a data frame returning the data frame
 without modification
>>>>>>> upstream/develop
}
