% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_tee}
\alias{skim_tee}
\title{Print useful summary statistic from a data frame without modification}
\usage{
skim_tee(.data)
}
\arguments{
\item{.data}{A tbl, or an object that can be coerced into a tbl.}
}
\value{
The input data frame.
}
\description{
Print useful summary statistic from a data frame without modification
}
