% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_tee}
\alias{skim_tee}
\title{Get useful summary statistic from a data frame and return the object}
\usage{
skim_tee(.data)
}
\arguments{
\item{.data}{A tbl, or an object that can be coerced into a tbl.}
}
\value{
A \code{skim_df} object, which can be treated like a
  tbl in most instances.
}
\description{
Get useful summary statistic from a data frame and return the object
}
