% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{skim_to_list}
\alias{skim_to_list}
\title{Skim results returned as a list}
\usage{
skim_to_list(.data, ...)
}
\arguments{
\item{.data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\value{
A list of data frames, one per variable type containing relevant
columns for the type
}
\description{
Skim results returned as a list
}
\examples{
skim_to_list(iris)
}
