% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_to_list}
\alias{skim_to_list}
\title{Create a list of skim tables instead of printing result}
\usage{
skim_to_list(x, ...)
}
\arguments{
\item{x}{A \code{dataframe}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of tibbles.
}
\description{
Returns a list of tibbles (also data frames) with one list element 
per data type. Each column contains the formatted values. 
This facilitates additional processing. 
Note that this is not pipeable.
}
\examples{
skim_to_list(iris)
iris \%>\% skim_to_list()

# Save the result
sl <- iris \%>\% skim_to_list() 
sl[["numeric"]]
kable(sl$numeric)

# Or grouped, this uses the magrittr exposition pipe
# see ?magrittr::`\%$\%`
library(magrittr)
iris \%>\%
  dplyr::group_by(Species) \%>\%
  skim_to_list() \%$\%
  kable(numeric)
}
