% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim_to_list}
\alias{skim_to_list}
\title{Print skim result and return a list of tibbles}
\usage{
skim_to_list(x, ...)
}
\arguments{
\item{x}{A \code{dataframe}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of tibbls.
}
\description{
Returns a list of tibbles with one list element 
 per data type. Each column contains the formatted values. 
 This facilitates additional processing. 
 Note that this is not pipeable.
}
\examples{
   skim_to_list(mtcars)
   mtcars \%>\% skim_to_list()
   l <- mtcars \%>\% skim_to_list() 
   l[["numeric"]]
}
