% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_v.R
\name{skim_v}
\alias{skim_v}
\title{Extract summary statistics for vector.}
\usage{
skim_v(x, vector_type = class(x))
}
\arguments{
\item{x}{A vector}

\item{vector_type}{A character vector that specifies which group of funs
to grab for summarizing.}
}
\value{
A tall tbl, containing the vector's name, type, potential levels
 and a series of summary statistics.
}
\description{
The set of functions used to summarize the vector depends on the vectors
type. See \code{\link{skim_with}} and
\code{\link{skim_format}} for how \code{skim_v} can be customized.
}
\keyword{internal}
