% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_v.R
\name{skim_v}
\alias{skim_v}
\title{Extract summary statistics for vector}
\usage{
skim_v(x, FUNS)
}
\arguments{
\item{x}{A vector}

\item{FUNS}{A list of functions to apply to x to compute summary statistics.
each function should return a numeric value.}
}
\value{
A tall tbl, containing the vector's name, type, potential levels
  and a series of summary statistics.
}
\description{
Extract summary statistics for vector
}
\keyword{internal}
