% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr-package.R
\docType{package}
\name{skimr-package}
\alias{skimr-package}
\alias{skimr}
\title{Skim a data frame}
\description{
This package provides an alternative to the default summary functions
within R. The package's API is tidy, functions take data frames, return
data frames and can work as part of a pipeline. The returned \code{skimr}
object is subsettable and offers a human readable output.
}
\details{
\code{skimr} is opinionated, providing a strong set of summary statistics
that are generated for a variety of different data types. It is also
provides an API for customization. Users can change both the functions
dispatched and the way the results are formatted.
}
