% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr-package.R
\docType{package}
\name{skimr-package}
\alias{skimr-package}
\alias{skimr}
\title{skimr
 It provides functionality to deal with summary statistics iteratively and interactively as part of a pipeline. The #'returned skimr object is subsettable and offers a human readable output.}
\description{
skimr
 It provides functionality to deal with summary statistics iteratively and interactively as part of a pipeline. The #'returned skimr object is subsettable and offers a human readable output.
}
\keyword{package}
