% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.skim_df}
\alias{summary.skim_df}
\title{Summary function for skim_df}
\usage{
\method{summary}{skim_df}(object, ...)
}
\arguments{
\item{object}{a skim dataframe.}

\item{...}{Additional arguments affecting the summary produced. Not used.}
}
\value{
A summary of the skim data frame.
}
\description{
This is a method of the generic function \code{\link[=summary]{summary()}}.
}
\examples{
\dontrun{
a <- skim(mtcars)
summary(a)
}
}
