% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{to_long}
\alias{to_long}
\title{Create "long" skim output}
\usage{
to_long(.data, ..., skim_fun = skim)
}
\arguments{
\item{.data}{A data frame or an object that can be coerced into a data frame.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\value{
A tibble
}
\description{
Skim results returned as a tidy long data frame with four columns:
variable, type, stat and formatted.
}
\examples{
to_long(iris)
to_long(skim(iris))
}
